refactor: Make `pl.duration` non-anonymous
