+++ crates/polars-plan/src/dsl/functions/temporal.rs
@@ -257,95 +257,7 @@ impl DurationArgs {
 /// Construct a column of [`Duration`] from the provided [`DurationArgs`]
 #[cfg(feature = "temporal")]
 pub fn duration(args: DurationArgs) -> Expr {
-    let function = SpecialEq::new(Arc::new(move |s: &mut [Series]| {
-        if s.iter().any(|s| s.is_empty()) {
-            return Ok(Some(Series::new_empty(
-                s[0].name(),
-                &DataType::Duration(args.time_unit),
-            )));
-        }
-
-        // TODO: Handle overflow for UInt64
-        let weeks = s[0].cast(&DataType::Int64).unwrap();
-        let days = s[1].cast(&DataType::Int64).unwrap();
-        let hours = s[2].cast(&DataType::Int64).unwrap();
-        let minutes = s[3].cast(&DataType::Int64).unwrap();
-        let seconds = s[4].cast(&DataType::Int64).unwrap();
-        let mut milliseconds = s[5].cast(&DataType::Int64).unwrap();
-        let mut microseconds = s[6].cast(&DataType::Int64).unwrap();
-        let mut nanoseconds = s[7].cast(&DataType::Int64).unwrap();
-
-        let is_scalar = |s: &Series| s.len() == 1;
-        let is_zero_scalar = |s: &Series| is_scalar(s) && s.get(0).unwrap() == AnyValue::Int64(0);
-
-        // Process subseconds
-        let max_len = s.iter().map(|s| s.len()).max().unwrap();
-        let mut duration = match args.time_unit {
-            TimeUnit::Microseconds => {
-                if is_scalar(&microseconds) {
-                    microseconds = microseconds.new_from_index(0, max_len);
-                }
-                if !is_zero_scalar(&nanoseconds) {
-                    microseconds = microseconds + (nanoseconds / 1_000);
-                }
-                if !is_zero_scalar(&milliseconds) {
-                    microseconds = microseconds + (milliseconds * 1_000);
-                }
-                microseconds
-            },
-            TimeUnit::Nanoseconds => {
-                if is_scalar(&nanoseconds) {
-                    nanoseconds = nanoseconds.new_from_index(0, max_len);
-                }
-                if !is_zero_scalar(&microseconds) {
-                    nanoseconds = nanoseconds + (microseconds * 1_000);
-                }
-                if !is_zero_scalar(&milliseconds) {
-                    nanoseconds = nanoseconds + (milliseconds * 1_000_000);
-                }
-                nanoseconds
-            },
-            TimeUnit::Milliseconds => {
-                if is_scalar(&milliseconds) {
-                    milliseconds = milliseconds.new_from_index(0, max_len);
-                }
-                if !is_zero_scalar(&nanoseconds) {
-                    milliseconds = milliseconds + (nanoseconds / 1_000_000);
-                }
-                if !is_zero_scalar(&microseconds) {
-                    milliseconds = milliseconds + (microseconds / 1_000);
-                }
-                milliseconds
-            },
-        };
-
-        // Process other duration specifiers
-        let multiplier = match args.time_unit {
-            TimeUnit::Nanoseconds => NANOSECONDS,
-            TimeUnit::Microseconds => MICROSECONDS,
-            TimeUnit::Milliseconds => MILLISECONDS,
-        };
-        if !is_zero_scalar(&seconds) {
-            duration = duration + seconds * multiplier;
-        }
-        if !is_zero_scalar(&minutes) {
-            duration = duration + minutes * (multiplier * 60);
-        }
-        if !is_zero_scalar(&hours) {
-            duration = duration + hours * (multiplier * 60 * 60);
-        }
-        if !is_zero_scalar(&days) {
-            duration = duration + days * (multiplier * SECONDS_IN_DAY);
-        }
-        if !is_zero_scalar(&weeks) {
-            duration = duration + weeks * (multiplier * SECONDS_IN_DAY * 7);
-        }
-
-        duration.cast(&DataType::Duration(args.time_unit)).map(Some)
-    }) as Arc<dyn SeriesUdf>);
-
-    // TODO: Make non-anonymous
-    Expr::AnonymousFunction {
+    Expr::Function {
         input: vec![
             args.weeks,
             args.days,
@@ -356,14 +268,11 @@ pub fn duration(args: DurationArgs) -> Expr {
             args.microseconds,
             args.nanoseconds,
         ],
-        function,
-        output_type: GetOutput::from_type(DataType::Duration(args.time_unit)),
+        function: FunctionExpr::TemporalExpr(TemporalFunction::Duration(args.time_unit)),
         options: FunctionOptions {
             collect_groups: ApplyOptions::ElementWise,
             input_wildcard_expansion: true,
-            fmt_str: "duration",
             ..Default::default()
         },
     }
-    .alias("duration")
 }
